+++ discord/flags.py
@@ -1586,6 +1586,14 @@ class ChannelFlags(BaseFlags):
         """
         return 1 << 15
 
+    @flag_value
+    def is_spoiler_channel(self):
+        """:class:`bool`: Returns ``True`` if the channel is a spoiler channel.
+
+        .. versionadded:: 2.9
+        """
+        return 1 << 21
+
 
 @fill_with_flags()
 class AttachmentFlags(BaseFlags):